Ruxin-package {StatComp20041}	R Documentation

This package provides several functions for ranking numbers from small to large.

Description

R codes for ranking numbers are provided in this package

Author(s)

Ruxin Shi

Maintainer: Ruxin Shi <shiruxin@msu.edu>

[Package StatComp20041 version 1.0 Index]